+++ src/main/java/io/greenscreens/foreign/Trivial.java
@@ -1,0 +1,19 @@
+/*
+ * Copyright (C) 2015, 2023 Green Screens Ltd.
+ */
+package io.greenscreens.foreign;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Annotation to identify foreign method is a trivial method, allowing JVM
+ * internal optimizations
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD})
+public @interface Trivial {
+
+}